+++ src/main/java/cmanager/util/DateTimeUtil.java
@@ -8,8 +8,16 @@ import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 
+/** Collection of utility methods for date and time handling. */
 public class DateTimeUtil {
 
+    /**
+     * Check if the given datetime is older than the given number of months.
+     *
+     * @param oldTime The datetime to check the invalidation for.
+     * @param maximumAgeInMonths The maximum age in months allowed.
+     * @return Whether the given datetime is too old.
+     */
     public static boolean isTooOldWithMonths(final LocalDateTime oldTime, int maximumAgeInMonths) {
         final LocalDateTime invalidationTime = oldTime.plusMonths(maximumAgeInMonths);
         final LocalDateTime now = LocalDateTime.now();
@@ -17,6 +25,13 @@ public class DateTimeUtil {
         return invalidationTime.isBefore(now);
     }
 
+    /**
+     * Check if the given file is older than the given number of months.
+     *
+     * @param file The file to check the modification time for.
+     * @param maximumAgeInMonths The maximum age in months allowed.
+     * @return Whether the file is too old.
+     */
     public static boolean isTooOldWithMonths(final File file, int maximumAgeInMonths) {
         final LocalDateTime fileModifiedDate =
                 LocalDateTime.ofInstant(
@@ -24,6 +39,13 @@ public class DateTimeUtil {
         return DateTimeUtil.isTooOldWithMonths(fileModifiedDate, maximumAgeInMonths);
     }
 
+    /**
+     * Check if the given file is too old.
+     *
+     * @param file The file to check the modification time for.
+     * @param expirationDatetime The datetime when the file should expire.
+     * @return Whether the file is too old.
+     */
     public static boolean isTooOld(final File file, final LocalDateTime expirationDatetime) {
         final LocalDateTime fileModifiedDate =
                 LocalDateTime.ofInstant(
@@ -65,5 +87,27 @@ public class DateTimeUtil {
         }
 
         return null;
+    }
+
+    /**
+     * Check if the given datetimes only differ by the given day offset. The offset is applied on
+     * both negative and positive direction to make the comparison independent of the parameter time
+     * order.
+     *
+     * @param dateTime1 The first datetime.
+     * @param dateTime2 The second datetime.
+     * @param dayOffset The maximum offset in days allowed.
+     * @return If the two datetimes only differ in the given number of days.
+     */
+    public static boolean isInDayRange(
+            final ZonedDateTime dateTime1, final ZonedDateTime dateTime2, final int dayOffset) {
+        if (dateTime1 == null || dateTime2 == null) {
+            return false;
+        }
+
+        final ZonedDateTime dateTime1Minus = dateTime1.minusDays(dayOffset);
+        final ZonedDateTime dateTime1Plus = dateTime1.plusDays(dayOffset);
+
+        return dateTime1Minus.isBefore(dateTime2) && dateTime2.isBefore(dateTime1Plus);
     }
 }